refactor: rename click-up related test naming
